Create README.md
Recommending preferred contents to TV enthusiasts for media companies such as Netflix, YouTube TV
or Amazon Prime is crucial. There is a need to automatically tag metadata for content on those 
platforms to help with data analysis and organization across their users and personalize
recommendations. The goal therefore is to train a neural network on a convolutional neural network 
(CNN) architecture to be able to predict celebrity given an image as an input.